♻️ rename specHelper to testSetupBuilder
