+++ packages/rum-core/src/boot/rumPublicApi.spec.ts
@@ -2,8 +2,8 @@ import type { RelativeTime, TimeStamp, Context } from '@datadog/browser-core'
 import { ONE_SECOND, getTimeStamp, display, DefaultPrivacyLevel } from '@datadog/browser-core'
 import { cleanupSyntheticsWorkerValues, mockSyntheticsWorkerValues } from '../../../core/test/syntheticsWorkerValues'
 import { initEventBridgeStub, deleteEventBridgeStub } from '../../../core/test/specHelper'
-import type { TestSetupBuilder } from '../../test/specHelper'
-import { setup } from '../../test/specHelper'
+import type { TestSetupBuilder } from '../../test/testSetupBuilder'
+import { setup } from '../../test/testSetupBuilder'
 import type { HybridInitConfiguration, RumInitConfiguration } from '../domain/configuration'
 import { ActionType } from '../rawRumEvent.types'
 import { noopRecorderApi } from '../../test/noopRecorderApi'
