+++ packages/rum-core/src/boot/startRum.spec.ts
@@ -12,8 +12,8 @@ import {
 import { createNewEvent, interceptRequests } from '../../../core/test/specHelper'
 import type { RumSessionManagerMock } from '../../test/mockRumSessionManager'
 import { createRumSessionManagerMock } from '../../test/mockRumSessionManager'
-import type { TestSetupBuilder } from '../../test/specHelper'
-import { setup } from '../../test/specHelper'
+import type { TestSetupBuilder } from '../../test/testSetupBuilder'
+import { setup } from '../../test/testSetupBuilder'
 import type { RumPerformanceNavigationTiming, RumPerformanceEntry } from '../browser/performanceCollection'
 import type { LifeCycle } from '../domain/lifeCycle'
 import { LifeCycleEventType } from '../domain/lifeCycle'
