+++ packages/rum-core/src/domain/contexts/urlContexts.spec.ts
@@ -1,7 +1,7 @@
 import type { RelativeTime } from '@datadog/browser-core'
 import { relativeToClocks } from '@datadog/browser-core'
-import type { TestSetupBuilder } from '../../../test/specHelper'
-import { setup } from '../../../test/specHelper'
+import type { TestSetupBuilder } from '../../../test/testSetupBuilder'
+import { setup } from '../../../test/testSetupBuilder'
 import { LifeCycleEventType } from '../lifeCycle'
 import type { ViewCreatedEvent, ViewEndedEvent } from '../rumEventsCollection/view/trackViews'
 import type { UrlContexts } from './urlContexts'
