+++ packages/rum-core/src/domain/rumEventsCollection/longTask/longTaskCollection.spec.ts
@@ -1,8 +1,8 @@
 import type { Duration, RelativeTime, ServerDuration } from '@datadog/browser-core'
 import type { RumSessionManagerMock } from '../../../../test/mockRumSessionManager'
 import { createRumSessionManagerMock } from '../../../../test/mockRumSessionManager'
-import type { TestSetupBuilder } from '../../../../test/specHelper'
-import { setup } from '../../../../test/specHelper'
+import type { TestSetupBuilder } from '../../../../test/testSetupBuilder'
+import { setup } from '../../../../test/testSetupBuilder'
 import type { RumPerformanceEntry, RumPerformanceLongTaskTiming } from '../../../browser/performanceCollection'
 import { RumEventType } from '../../../rawRumEvent.types'
 import { LifeCycleEventType } from '../../lifeCycle'
