+++ packages/rum-core/src/domain/rumEventsCollection/resource/resourceCollection.spec.ts
@@ -9,8 +9,8 @@ import {
 } from '@datadog/browser-core'
 import type { RumFetchResourceEventDomainContext } from '../../../domainContext.types'
 import { createResourceEntry } from '../../../../test/fixtures'
-import type { TestSetupBuilder } from '../../../../test/specHelper'
-import { setup } from '../../../../test/specHelper'
+import type { TestSetupBuilder } from '../../../../test/testSetupBuilder'
+import { setup } from '../../../../test/testSetupBuilder'
 import type { RawRumResourceEvent } from '../../../rawRumEvent.types'
 import { RumEventType } from '../../../rawRumEvent.types'
 import { LifeCycleEventType } from '../../lifeCycle'
