+++ packages/rum-core/src/domain/rumEventsCollection/view/trackViewMetrics.spec.ts
@@ -1,8 +1,8 @@
 import type { Context, RelativeTime, Duration } from '@datadog/browser-core'
 import { addDuration, relativeNow } from '@datadog/browser-core'
 import type { RumEvent } from '../../../rumEvent.types'
-import type { TestSetupBuilder } from '../../../../test/specHelper'
-import { setup } from '../../../../test/specHelper'
+import type { TestSetupBuilder } from '../../../../test/testSetupBuilder'
+import { setup } from '../../../../test/testSetupBuilder'
 import type { RumPerformanceNavigationTiming } from '../../../browser/performanceCollection'
 import { FrustrationType, RumEventType } from '../../../rawRumEvent.types'
 import type { LifeCycle } from '../../lifeCycle'
