+++ packages/rum-core/src/domain/rumEventsCollection/view/trackViews.spec.ts
@@ -7,8 +7,8 @@ import {
   relativeNow,
   resetExperimentalFeatures,
 } from '@datadog/browser-core'
-import type { TestSetupBuilder } from '../../../../test/specHelper'
-import { setup } from '../../../../test/specHelper'
+import type { TestSetupBuilder } from '../../../../test/testSetupBuilder'
+import { setup } from '../../../../test/testSetupBuilder'
 import type {
   RumLargestContentfulPaintTiming,
   RumPerformanceNavigationTiming,
