+++ packages/rum-core/src/domain/rumEventsCollection/view/viewCollection.spec.ts
@@ -1,7 +1,7 @@
 import type { Duration, RelativeTime, ServerDuration, TimeStamp } from '@datadog/browser-core'
 import type { RecorderApi } from '../../../boot/rumPublicApi'
-import type { TestSetupBuilder } from '../../../../test/specHelper'
-import { setup } from '../../../../test/specHelper'
+import type { TestSetupBuilder } from '../../../../test/testSetupBuilder'
+import { setup } from '../../../../test/testSetupBuilder'
 import type { RawRumViewEvent } from '../../../rawRumEvent.types'
 import { RumEventType, ViewLoadingType } from '../../../rawRumEvent.types'
 import { LifeCycleEventType } from '../../lifeCycle'
