+++ packages/rum-core/src/domain/startCustomerDataTelemetry.spec.ts
@@ -1,7 +1,7 @@
 import type { BatchFlushEvent, Context, ContextManager, TelemetryEvent } from '@datadog/browser-core'
 import { resetExperimentalFeatures, TelemetryService, startTelemetry, Observable } from '@datadog/browser-core'
-import type { TestSetupBuilder } from '../../test/specHelper'
-import { setup } from '../../test/specHelper'
+import type { TestSetupBuilder } from '../../test/testSetupBuilder'
+import { setup } from '../../test/testSetupBuilder'
 import { RumEventType } from '../rawRumEvent.types'
 import type { RumEvent } from '../rumEvent.types'
 import type { FeatureFlagContexts } from './contexts/featureFlagContext'
