+++ packages/rum/src/boot/recorderApi.spec.ts
@@ -4,8 +4,8 @@ import { LifeCycleEventType } from '@datadog/browser-rum-core'
 import { createNewEvent, deleteEventBridgeStub, initEventBridgeStub } from '../../../core/test/specHelper'
 import type { RumSessionManagerMock } from '../../../rum-core/test/mockRumSessionManager'
 import { createRumSessionManagerMock } from '../../../rum-core/test/mockRumSessionManager'
-import type { TestSetupBuilder } from '../../../rum-core/test/specHelper'
-import { setup } from '../../../rum-core/test/specHelper'
+import type { TestSetupBuilder } from '../../../rum-core/test/testSetupBuilder'
+import { setup } from '../../../rum-core/test/testSetupBuilder'
 import type { DeflateWorker, startDeflateWorker } from '../domain/segmentCollection'
 import type { StartRecording } from './recorderApi'
 import { makeRecorderApi } from './recorderApi'
